Rename capybara helpers
