+++ spec/rails_helper.rb
@@ -27,7 +27,7 @@ require 'alchemy/test_support/shared_uploader_examples'
 require_relative 'factories'
 require_relative "support/hint_examples.rb"
 require_relative "support/transformation_examples.rb"
-require_relative "support/capybara_select2.rb"
+require_relative "support/capybara_helpers.rb"
 require_relative 'support/custom_news_elements_finder'
 
 ActionMailer::Base.delivery_method = :test
